buy_books/views.py
from django.shortcuts import render
from rest_framework import generics

from .models import Post
from .serializer import PostSerializer

class PostsList(generics.ListCreateAPIView):
    queryset = Post.objects.all()
    serializer_class = PostSerializer

class PostDetails(generics.RetrieveUpdateDestroyAPIView):
    queryset = Post.objects.all()
    serializer_class = PostSerializer